requirements.txt
anyio==3.6.2
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mysqlclient==2.1.1
pycodestyle==2.10.0
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
